refactor: seperate authorization from authentication
Function which was previously titled 'authenticate' in file 'verifyToken.js' implemented an authorization functionality which checked if the 'req.params.id' matched the 'authData.user._id'. While this works fine for user endpoints which use the user id, this wont work for endpoints based on other data types. To fix this, the authorization functionality was extracted from the function and placed in the route controllers, where the controller can handle this based on the unique demands of its routes data type and endpoint. Due to this change, what was previously title 'authenticate' was renamed to 'verifyToken' to reflect the more focused nature of the function. Other names were also altered to reflect this change.